Move API sections into individual files
in order to improve maintainability as the main file was bloated and it was difficult to read
