+++ rocketchat_API/APISections/permissions.py
@@ -1,0 +1,7 @@
+from rocketchat_API.APISections.base import RocketChatBase
+
+
+class RocketChatPermissions(RocketChatBase):
+    def permissions_list_all(self, **kwargs):
+        """Returns all permissions from the server."""
+        return self.call_api_get("permissions.listAll", kwargs=kwargs)
